requirements.txt
openai
elevenlabs==0.2.27
spacy
nltk
SpeechRecognition
pyaudio
python-dotenv
wikipedia
requests
apscheduler
duckduckgo-search
fastapi
uvicorn
websockets
scikit-learn
numpy
chromadb
pvporcupine
colorama
pyyaml
jinja2
python-multipart
